+++ client/src/components/HomeFeed/HomeFeed.jsx
@@ -3,91 +3,38 @@ import React, { useState, useEffect } from 'react';
 import SearchUsernames from './SearchUsernames.jsx';
 import { Link } from 'react-router-dom'
 import TodaysGoals from '../myProfile/todaysGoal.jsx';
+import Rankings from './Rankings.jsx';
+import PersonalRank from './PersonalRank.jsx';
 
 
 function HomeFeed() {
 
-  const [friends, setFriends] = useState([]);
   const [goals, setGoals] = useState({});
+  const [id, setId] = useState(1);
 
   useEffect(() => {
-    axios.get('/rankings')
-      .then(response => response.data)
-      .then(result => setFriends(result[0].friends))
-      .catch(error => error)
     axios.get('/todaysgoals')
       .then(results => {
-          setGoals({
-            calories: results.data.calories,
-            water: results.data.water,
-            weight: results.data.weight
-          })
+        setGoals({
+          calories: results.data.calories,
+          water: results.data.water,
+          weight: results.data.weight
         })
+      })
   }, []);
-
-  console.log('goals', goals)
-
-  sortFriends();
-
-  function sortFriends() {
-    friends.forEach(friend => {
-      // returns negative if they missed the goal
-      let water = Math.abs(100 - (friend['goals'][0]['wateraverage'] * 100))
-      // returns negative if goal is exceeded
-      let calories = Math.abs(100 - (friend['goals'][0]['caloriesaverage'] * 100))
-      let calculate = water + calories;
-      friend['sorting'] = calculate.toFixed(2);
-    })
-  }
 
   return (
     <div id='home-page'>
       {/* Home Feed Search */}
-
       <SearchUsernames />
-
-      <div className='home-placeholder'>
+      <h4>Your Daily Status:</h4>
+      <div className='home-daily-stats'>
         <TodaysGoals goals={goals} />
+        <PersonalRank />
       </div>
-
-      <h4>Your Friend's Rankings: </h4>
-      {/* Friend's List Tile */}
-      {friends.filter(friend => friend.sorting >= 0)
-        .sort((a, b) => a.sorting - b.sorting)
-        .concat(friends
-          .filter(friend => friend.sorting < 0)
-          .sort((a, b) => b.sorting - a.sorting))
-        .map((friend, index) => (
-          <div
-            className='pic-tile-friend-tile'
-            key={index}
-            onClick={() => console.log('On click needs to route to', friend.friendfirst)}
-          >
-            <div className='pic-tile-ranking'>
-              #{index + 1}
-            </div>
-            {/* Profile Picture */}
-            <div className='pic-tile-friend-left-pic' style={{ 'width': '15%', }}>
-              <img style={{ 'maxHeight': '50px' }} src={friend.profilephoto}></img>
-            </div>
-            {/* Friend Information */}
-            <div
-              className='pic-tile-friend-right-info'
-              style={{ 'fontSize': '14px', 'width': '85%' }}
-            >
-              <b>{friend.friendusername}:</b>
-              <li>
-                {Math.round(friend.goals[0].wateraverage * 100)}% of my water goal.
-              </li>
-              <li>
-                {Math.round(friend.goals[0].caloriesaverage * 100)}% of my calories goal.
-              </li>
-            </div>
-          </div>
-        ))}
+      <Rankings id={id}/>
       <div className='feed-bottom'></div>
     </div>
-
   );
 }
