+++ server/server.js
@@ -44,7 +44,6 @@ app.get('/user', (req, res) => {
 
 // get current user's friends
 app.get('/friends', (req, res) => {
-  console.log('-', req.query.friendId)
   let friendId = req.query.friendId;
   db.client.query(`
       select id, firstname, lastname, picture, descriptionmessage,
@@ -175,10 +174,35 @@ app.put('/updateToday', (req, res) => {
 })
 
 // get home feed rankings data
-app.get('/rankings', (req, res) => {
-  let friendId = 1;
+app.get(`/rankings`, (req, res) => {
+  let friendId = req.query.friendId;
   db.client.query(`
     select id, firstname, lastname, picture, descriptionmessage,
+
+    (
+      select array_to_json(array_agg(row_to_json(d)))
+      from (
+        select watergoal, caloriegoal, weightgoal,
+        (
+          select avg(water)/watergoal
+          from dailyData
+          where dailyData.userId = 1
+        ) as wateraverage ,
+        (
+          select avg(calories)/caloriegoal
+          from dailyData
+          where dailyData.userId = 1
+        ) as caloriesaverage ,
+        (
+          select avg(weight)/weightgoal
+          from dailyData
+          where dailyData.userId = 1
+        ) as weightaverage
+        from goals
+        where goals.userId = 1 
+      ) d
+    ) as userdata,
+
       (
         select array_to_json(array_agg(row_to_json(d)))
         from (
@@ -223,25 +247,26 @@ app.get('/rankings', (req, res) => {
                 where dailyData.userId = friendId
               ) as weightaverage
               from goals
-              where goals.userId = friendId
+              where goals.userId = friendId 
             ) d
           ) as goals
           from friends
-          where friends.userID = users.id
+          where friends.userID = users.id 
         ) d
       ) as friends
     from users
     where users.id = ${friendId}
   `, (err, data) => {
     if (err) {
-      // console.log('error from server -', err)
+      console.log('error from server -', err)
       res.send(err);
     } else {
-      // console.log('rows from server /rankings - ', data.rows)
+      console.log('rows from server /rankings - ', data.rows)
       res.send(data.rows);
     }
   })
 });
+
 // get friend profile information
 // SELECT * FROM dailyData
 // WHERE userID = ${friendId} AND shareBoolean=true
@@ -303,6 +328,7 @@ app.get('/friendProfile', (req, res) => {
       }
     })
 });
+
 app.post('/addfriend', (req, res) => {
   let friendId = 7;
   let userId = 1;
@@ -321,6 +347,7 @@ app.post('/addfriend', (req, res) => {
     }
   })
 });
+
 app.delete('/removefriend', (req, res) => {
   let friendId = 7;
   let userId = 1;
